Add files via upload
Updated the html to reflect newer XHTML code. Broken out the javascript code to show the working if statement and looped the cycle of images. In reference to the ++index vs index++ in the particular instance of how it was coded previously it did not really affect the end result based on whether it incremented before or after since it was a pretty generic image swap. However, I am aware that it would be an issue if you looking to code with a more specific swap in mind. I do hope that this covers everything that you were looking for. I wasn't sure if you wanted me to include it, but all the files have been validated and checked. Specifically the html and css were validated by W3c and the js by beautify tools